db now supports multiple classes
