+++ package.json
@@ -20,7 +20,6 @@
     "next": "13.5.4",
     "react": "^18",
     "react-dom": "^18",
-    "react-simple-maps": "^3.0.0",
     "socket.io": "^4.7.2",
     "socket.io-client": "^4.7.2",
     "sqlite3": "^5.1.6",
