+++ src/components/Footer.js
@@ -1,0 +1,11 @@
+const Footer = () => {
+    return (
+        <footer className="bg-black text-white py-4 sticky bottom-0">
+            <div className="container mx-auto text-center">
+                <p>Released under the MIT License. Created by Alex Hadley, Nikit Patel and Gabriela Djulabic.</p>
+            </div>
+        </footer>
+    );
+  };
+  
+  export default Footer;
